Support FLAG_DELETED in Office365
When a local message is tagged deleted then update the cloud flags and
delete the message.

Using the Trash flag in a mail dir allows the cloud upload step to see the
original message and then upload the local flags before deleting it. For
instance, this will allow the Replied flag to be set upon deletion.

Once deleted in the cloud the next sync iteration will delete the message
locally.

Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>