gmail: Fix determining the start_history_id
Occasionally gmail returns history records that have junk in them, for
instance messages IDs that don't exist. This appears to be a bug.

Using the latest history ID seems to guarantee that a subsequent history
query will be empty and we can then avoid the junk and start again.

Go back to full query if delta fails for any reason.

Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>