Add cms-oauth
This is a command line program to get the OAUTH tokens from the credential
server. It is intended to fit into the 'call a program to get the token'
methodology that several tools are implementing.

Several options are provided to format the token and a built in SMTP
protocol tests that the server is working properly.

Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>