Manage concurrency to avoid running out of file descriptors
AIO does not seem to strictly pipeline work, so it is possible for a very
large number of message download tasks to be scheduled. Each one that gets
scheduled opens a file descriptor, and so we can run out of them fairly
fast.

Strictly limit the number of open files per mailbox during message
downloading.

Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>