Run changed_event.set() in main loop
Watchdog runs in a separate thread. When calling changed_event.set(),
from that separate thread, there is a "RuntimeError: Non-thread-safe
operation invoked on an event loop other than the current one". To
correct this, asyncio.run_coroutine_threadsafe is used to set the flag
in the main cfg.loop. This change results in immediate response from the
main.py when changes are made to watched folders - as should be the
case, rather than with a 2-3 seconds delay previously when the thread
unsafe code was used.

Also, redundant import of Observer is removed.